# style: Improve logging similarity to InfluxDB 